dags/bees/openbrewery/requirements.txt
pyspark==3.5.0 
delta-spark==3.1.0
apache-airflow==3.0.6
apache-airflow-providers-amazon==9.24.0
